Fix bug on confirm password keyup
There was a bug where if the passwords did not match and the user re-
enetered a password less than 8 characters. The password error field
would show the new error message saying less than 8 characters. If the
user the entered a password in the confirm password field that matched
the password that was not long enough both incorrect validation messages
would be removed and the inputs would go back to normal color.